Turn on testing for JSON under Mariadb 10.2.7 and greater
MariaDB adds a JSON alias for the LONGTEXT datatype and
all the same json functions work against it in any case.
What doesn't work is reflection, since it's just an alias,
and also the CAST we were using in one of our tests doesn't seem
to work for JSON which is probably also because it's not
a real datatype.

Change-Id: I44e5503db29ca2f04de8e521527978f34675a5e0